Clova-FinAgent/requirements.txt
# FastAPI 관련
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# LangChain 관련
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.20
langchain-naver>=0.0.1

# 기타 필수 패키지
python-dotenv>=1.0.0
typing-extensions>=4.8.0
requests>=2.31.0
aiofiles>=23.2.1

# 데이터베이스 관련
# sqlite3

# 기타 유틸리티
python-multipart>=0.0.6 

# 그래프
matplotlib
koreanize-matplotlib